test: add async test
