+++ collab-database/tests/database_test/helper.rs
@@ -16,6 +16,7 @@ use tempfile::TempDir;
 use tracing_subscriber::{fmt::Subscriber, util::SubscriberInitExt, EnvFilter};
 
 pub use crate::helper::*;
+
 pub struct DatabaseTest {
   database: Database,
 
@@ -64,19 +65,7 @@ pub fn create_database(uid: i64, database_id: &str) -> DatabaseTest {
 }
 
 pub fn create_database_with_db(uid: i64, database_id: &str) -> (Arc<CollabKV>, DatabaseTest) {
-  static START: Once = Once::new();
-  START.call_once(|| {
-    std::env::set_var("RUST_LOG", "collab_persistence=trace");
-    let subscriber = Subscriber::builder()
-      .with_env_filter(EnvFilter::from_default_env())
-      .with_ansi(true)
-      .finish();
-    subscriber.try_init().unwrap();
-  });
-
-  let tempdir = TempDir::new().unwrap();
-  let path = tempdir.into_path();
-  let db = Arc::new(CollabKV::open(path).unwrap());
+  let db = make_kv_db();
   let disk_plugin = CollabDiskPlugin::new(uid, db.clone()).unwrap();
   let snapshot_plugin = CollabSnapshotPlugin::new(uid, db.clone(), 5).unwrap();
