+++ collab-database/tests/user_test/script_test/script.rs
@@ -1,5 +1,4 @@
-use crate::helper::TestTextCell;
-use crate::user_test::helper::make_kv_db;
+use crate::helper::{make_kv_db, TestTextCell};
 use collab_database::block::CreateRowParams;
 use collab_database::database::DuplicatedDatabase;
 use collab_database::fields::Field;
